pyproject.toml
# ===== PROJECT ===========================================================
#
[project]
name = 'pfore-cloud-utilities'
authors = [
    {name = 'BahaEddine Abrougui', email = 'bahaeddine.abrougui@lidl.com'},
]
description = 'Provides utility functions for cloud-based workflows.'
readme = 'README.md'
keywords = [
    'databricks',
    'azure',
    'mlops',
]
classifiers = [
    'Development Status :: 1 - Planning',
    'Intended Audience :: Science/Research',
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    # When specified, pypip upload will be rejected
    'Private :: Do Not Upload',
]
requires-python = '>=3.9'
dependencies = [
    'databricks-sdk >=0.9.0, <1.0.0',
    'numpy >=1.26.0, <1.27.0',
    'pandas >=2.1.1, <2.2.0',
    'pyarrow >= 13.0.0, <14.0.0',
    'azure-identity >=1.14.0, <1.15.0',
    'azure-storage-blob >=12.18.2, <12.19.0',
]
# Get version dynamically from env variable ARTIFACT_LABEL
dynamic = ['version']

[project.urls]
repository = 'https://dev.azure.com/schwarzit/schwarzit.uapc-pfore/_git/cloud-utilities/'
documentation = 'https://dev.azure.com/schwarzit/schwarzit.uapc-pfore/_git/cloud-utilities?path=/README.md'


# ===== BUILD-SYSTEM ======================================================
#
# Requirements for building the Python package
[build-system]
requires = ['setuptools>=45']
build-backend = 'setuptools.build_meta'


# ----- setuptools --------------------------------------------------------
#
# Find all (sub-)modules of the Python package
[tool.setuptools.packages.find]